Annotate 1-async_comprehension properly

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -6,6 +6,6 @@ from typing import List
 agen = __import__("0-async_generator").async_generator
 
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """Async comprehend an async generator"""
     return [i async for i in agen()]
